+++ CHANGELOG.md
@@ -9,6 +9,7 @@ This project <em>does not yet</em> adheres to [Semantic Versioning](https://semv
 
 ### Added
 - Taxon determination, citations and collecting event information in specimen record on browse OTU
+- Serial facet on filter sources
 
 ## [0.14.1] - 2020-10-22